fix server dependencies + allow standalone
We only need to depend on the database, if one is actually used. the
database_type var defaults to postgres, so we can't only rely on that.
there is also a manage_database var. This allows us to only provision
the zabbix-server on one node, and the database on a different one. We
need to configure the systemd unit correctly for this.